some .git cleanup

+++ fish/install.sh
@@ -1,4 +1,4 @@
-#/bin/bash
+#!/bin/bash
 
 if test $(which fish) ; then
 	echo fish already installed
@@ -9,6 +9,13 @@ else
 
 fi
 
+# if file is already symbolic link
+if [[ -L ~/.config/fish/config.fish ]] ; then 
+	echo "Fish Config already installed"
+else 
 	ln -s ~/.dotfiles/fish/config.fish ~/.config/fish/config.fish
+fi
 
+curl https://git.io/fisher --create-dirs -sLo ~/.config/fish/functions/fisher.fish
 
+fish -c "fisher add franciscolourenco/done"
